@@ -1,9 +1,8 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 import App from './App.vue'
+import createRouter from './router'
 import router from './router'
 
-Vue.config.productionTip = false
-
-new Vue({
-	render: h => h(App),
-}).$mount('#app')
+const app = createApp(App).use(router)
+app.use(createRouter)
+app.mount('#app')
