@@ -1,23 +1,26 @@
 <template>
-	<div id="app">
-		<img src="./assets/logo.png" alt="">
-		<router-view/>
-	</div>
+  <router-view/>
 </template>
 
 <script>
-	export default {
-		name: 'app'
-	}
+export default {
+  name: 'app'
+}
 </script>
 
 <style>
-	#app {
-		font-family: 'Avenir', Helvetica, Arial, sans-serif;
-		-webkit-font-smoothing: antialiased;
-		-moz-osx-font-smoothing: grayscale;
-		text-align: center;
-		color: #2c3e50;
-		margin-top: 60px;
-	}
+
+#app {
+  font-family: 'Avenir', Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+}
+
+body {
+  margin: 0;
+}
+
+@import 'styles/zcm-main.css';
+
 </style>
